Add a routes to GET /examples
related #5, #7.
