+++ db/mysql/my.conf
@@ -6,7 +6,11 @@ ft-min-word-len = 2
 innodb-ft-min-token-size = 2
 innodb-ft-enable-stopword = OFF
 
-character-set-server=utf8
+character-set-server = utf8mb4
+collation-server = utf8mb4_bin
+
+[mysql]
+default-character-set = utf8mb4
 
 [client]
-default-character-set=utf8
+default-character-set=utf8mb4
