+++ examples/admin.py
