+++ examples/tests.py
