+++ pyproject.toml
@@ -9,6 +9,8 @@ license = "MIT"
 python = "^3.8"
 mysqlclient = "^1.4.6"
 django = "^3.0.6"
+django-extensions = "^2.2.9"
+ipython = "^7.14.0"
 
 [tool.poetry.dev-dependencies]
